Create source code
The above example is a web automation script. It launches a website (randomwebsite.com, you can add your preferred site), finds the ‘Login’ (or Sign Up) element, and clicks on it. After that, it enters the credentials in the login page, clicks the login button, and redirects you to the website’s homepage.